// src/types/Cow.ts
export type Cow = {
  id: string;
  name: string;
  gender: string;
  birthDate: string;
  father: string;
  mothersFather: string;
  mothersGrandfather: string;
  grandmothersGrandfather: string;
  motherName: string;
  motherRegistrationNumber: string;
  motherScore: string;
  motherBirthDate: string;
  producerName: string;
  status: string;
  notes: string;
};